openpos-service/src/test/java/org/jumpmind/pos/service/EndpointDispatchInvocationHandlerTest.java
package org.jumpmind.pos.service;

import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
import static com.github.tomakehurst.wiremock.client.WireMock.get;
import static com.github.tomakehurst.wiremock.client.WireMock.post;
import static com.github.tomakehurst.wiremock.client.WireMock.put;
import static com.github.tomakehurst.wiremock.client.WireMock.status;
import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;

import java.math.BigDecimal;

import org.jumpmind.pos.util.model.ErrorResult;
import org.jumpmind.pos.util.web.ConfiguredRestTemplate;
import org.junit.Ignore;
import org.junit.Rule;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;

import com.fasterxml.jackson.databind.ObjectMapper;
import com.github.tomakehurst.wiremock.common.ConsoleNotifier;
import com.github.tomakehurst.wiremock.junit.WireMockRule;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = { TestServiceConfig.class })
public class EndpointDispatchInvocationHandlerTest {

    @Rule
    public WireMockRule wireMockRule = new WireMockRule(options().notifier(new ConsoleNotifier(true)));

    @Autowired
    EndpointDispatchInvocationHandler dispatcher;

    @Autowired
    TestEndpoint endpoint;

    @Autowired
    TestEndpointOverride override;

    ObjectMapper mapper = new ConfiguredRestTemplate().getMapper();
    
    EndpointDispatchInvocationHandler handler = new EndpointDispatchInvocationHandler();

    @Test
    public void testThatOverrideIsCalled() throws Throwable {
        assertEquals(0, endpoint.invokeCount);
        assertEquals(0, override.invokeCount);
        dispatcher.invoke(null, ITest.class.getMethod("test"), null);
        assertEquals(0, endpoint.invokeCount);
        assertEquals(1, override.invokeCount);
    }

    @Test
    public void testInvokeRemotePostWithResponseNoRequest() throws Throwable {
        stubFor(post(urlEqualTo("/check/deviceid/test001/version")).willReturn(status(200).withHeader("Content-Type", "application/json")
                .withBody(mapper.writeValueAsString(new TestResponse(new BigDecimal("1.11"), "abcd")))));

        TestResponse response = (TestResponse) handler.invokeRemote(config(), ITestService.class.getMethod("testPost", String.class),
                new Object[] { "test001" });

        assertNotNull(response);
        assertEquals(new BigDecimal("1.11"), response.total);
        assertEquals("abcd", response.message);
    }

    @Test
    public void testInvokeRemotePutWithRequestWithResponse() throws Throwable {
        stubFor(put(urlEqualTo("/check/deviceid/test001/yada")).willReturn(status(200).withHeader("Content-Type", "application/json")
                .withBody(mapper.writeValueAsString(new TestResponse(new BigDecimal("3.14"), "xyz")))));

        TestResponse response = (TestResponse) handler.invokeRemote(config(), ITestService.class.getMethod("testPut", String.class, TestRequest.class),
                new Object[] { "test001", new TestRequest("one", 1) });

        assertNotNull(response);
        assertEquals(new BigDecimal("3.14"), response.total);
        assertEquals("xyz", response.message);
    }

    @Test
    public void testInvokeRemotePutWithNoResponse() throws Throwable {
        stubFor(put(urlEqualTo("/check/deviceid/test001/nuttin")).willReturn(status(200)));

        handler.invokeRemote(config(), ITestService.class.getMethod("testPutNuttin", String.class, TestRequest.class),
                new Object[] { "test001", new TestRequest("one", 1) });

    }

    @Ignore("I can't make wire mock send 500 and the response body like spring does.  not sure what i'm doing wrong")
    @Test
    public void testInvokeRemotePutWithError() throws Throwable {
        ErrorResult result = new ErrorResult("this was a test", new NullPointerException());
        stubFor(put(urlEqualTo("/check/deviceid/test001/nuttin")).willReturn(aResponse().withHeader("Content-Type", "application/json")
                .withBody(mapper.writeValueAsString(result)).withStatus(501)));
        
        handler.invokeRemote(config(), ITestService.class.getMethod("testPutNuttin", String.class, TestRequest.class),
                new Object[] { "test001", new TestRequest("one", 1) });

    }

    @Test
    public void testInvokeRemoteGet() throws Throwable {
        stubFor(get(urlEqualTo("/check/getmesomeofthat")).willReturn(status(200).withHeader("Content-Type", "application/json")
                .withBody(mapper.writeValueAsString(new TestResponse(new BigDecimal("3.14"), "xyz")))));

        TestResponse response = (TestResponse) handler.invokeRemote(config(), ITestService.class.getMethod("testGet"), null);

        assertNotNull(response);
        assertEquals(new BigDecimal("3.14"), response.total);
        assertEquals("xyz", response.message);
    }

    private ServiceSpecificConfig config() {
        ServiceSpecificConfig config = new ServiceSpecificConfig();
        config.setHttpTimeout(30);
        config.setUrl("http://localhost:8080");
        return config;
    }

    static class TestRequest {
        String deviceId;
        int sequenceNumber;

        public TestRequest() {
        }

        public TestRequest(String deviceId, int sequenceNumber) {
            super();
            this.deviceId = deviceId;
            this.sequenceNumber = sequenceNumber;
        }

        public String getDeviceId() {
            return deviceId;
        }

        public void setDeviceId(String deviceId) {
            this.deviceId = deviceId;
        }

        public int getSequenceNumber() {
            return sequenceNumber;
        }

        public void setSequenceNumber(int sequenceNumber) {
            this.sequenceNumber = sequenceNumber;
        }

    }

    static class TestResponse {
        BigDecimal total;
        String message;

        public TestResponse() {
        }

        public TestResponse(BigDecimal total, String message) {
            super();
            this.total = total;
            this.message = message;
        }

        public BigDecimal getTotal() {
            return total;
        }

        public void setTotal(BigDecimal total) {
            this.total = total;
        }

        public String getMessage() {
            return message;
        }

        public void setMessage(String message) {
            this.message = message;
        }

    }

    @RestController("test")
    @RequestMapping("/check")
    interface ITestService {

        @RequestMapping(path = "/deviceid/{deviceid}/version", method = RequestMethod.POST)
        public TestResponse testPost(@PathVariable("deviceid") String deviceId);
        
        @RequestMapping(path = "/deviceid/{deviceid}/yada", method = RequestMethod.PUT)
        public TestResponse testPut(@PathVariable("deviceid") String deviceId, @RequestBody TestRequest request);
        
        @RequestMapping(path = "/deviceid/{deviceid}/nuttin", method = RequestMethod.PUT)
        public void testPutNuttin(@PathVariable("deviceid") String deviceId, @RequestBody TestRequest request);
        
        @RequestMapping(path = "/getmesomeofthat", method = RequestMethod.GET)
        public TestResponse testGet();
    }

    @RestController("test")
    @RequestMapping("/this/is/a/test")
    interface ITest {
        public void test();
    }

}
